Create ICEBLKv3
Firstly, you will need to connect to the Hive blockchain using a library such as dsteem or hive-js. This will allow you to listen for transactions and transfers related to the ICEBLK token and the "@icebrk" account.

Once you have connected to the blockchain, you can use the onCustomJson method to listen for any custom JSON transactions related to the "@icebrk" account. You can filter for transactions that involve the ICEBLK token and check if they are sell orders. If they are, you can then use the getMarketBuyAmountsOut method from the Tribaldex API to calculate the amount of BRK tokens you can buy with half of the SWAP.HIVE received from the sale.

Next, you can multiply the amount of BRK tokens purchased by 2 to get the RedeemValue and store it as a variable. You can then mark the timestamp of the ICEBLK sale, add 72 hours to it, and denote this as the "Melt Time". You can use the transfer method from the dsteem or hive-js library to send a transfer of 0.001 SWAP.HIVE to the user who bought the ICEBLK token/s with a memo that includes the RedeemValue and the "Melt Time".

To listen for incoming transfers of ICEBLK tokens to the "@icebrk" account, you can use the onTransfer method from the dsteem or hive-js library. When a transfer is received, you can check to make sure that 72 hours has elapsed from the time of purchase ("Melt Time" has been met) and transfer the RedeemValue in BRK tokens to the user according to the memo in the 0.001 SWAP.HIVE transfer which was sent prior.

You can use the Tribaldex API to get the current price of the BRK token and calculate the appropriate amount of BRK tokens to send to the user.

Overall, this task involves interacting with the Hive blockchain, the Tribaldex API, and the dsteem or hive-js library to listen for transactions and transfers, perform calculations, and send transfers. It will require a good understanding of Javascript and blockchain development.