Update ICEBLKv4
This function listens for incoming transfers of ICEBLK tokens to the "@icebrk" account, checks to make sure that 72 hours have elapsed from the time of purchase ("Melt Time" has been met) and transfers the RedeemValue in BRK tokens to the user according to the memo in the 0.001 SWAP.HIVE transfer which was sent prior. If the melt time has not been met, the function will log a message to the console and continue processing the next transaction. If the melt time has been met, the function will buy BRK tokens using half of the SWAP.HIVE received from the ICEBLK sale, transfer the appropriate amount of BRK tokens to the user according to the memo, and log messages to the console to indicate success or failure.